+++ packages/internal-utils/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./assert";
 export * from "./authIdentity";
 export * from "./prepareBroadcastRequest";
 export * from "./tokenUtils";
+export * from "./near";