+++ packages/crosschain-assetid/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @defuse-protocol/crosschain-assetid
 
+## 1.0.1
+
+### Patch Changes
+
+- e5d9a9c: Do not bundle into single file. Use `tsdown` for transpiling.
+
 ## 1.0.0
 
 ### Major Changes